Adds a File > Open Dialog
I'm not sure if we actually need it but at some point it may be useful. I was trying to learn async bevy in my last hour at work. This commit introduces a simple File > Open Dialog that should theoretically run in both WASM and natively.

## Notes

You need to filter eventa using a seperate system that handles the completion of your Task. Currently this system waits on any task that has a `Option<Vec<u8>>` response but ideally every async request should have a unique type associated with it so the system which handles the result will wait on the unique callback.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>