Switch to stable Rust version, rather than a pegged nightly
Back in April there was a small regression in rust-nightly that
required us to peg to nightly from a few days before. At that
time, we needed to use rust-nightly because of some leading-edge
features required by Bevy. This is no longer the case, and I
think we'll benefit from returning to the stable version. As
another benefit, this will no longer require manual installation
of a particular nightly toolchain for the WASM build.

Signed-off-by: Morgan Quigley <morgan@osrfoundation.org>