Adds a simulation preview button
As part of my push towards visuallization of a crowd in rmf_sandbox, I am adding a "simulation preview" button. This button currently does nothing, however it enables a `SimulationState` resource. This will be needed when a user wants to preview the crowd simulations prior to running it in gazebo.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>