Adds an API for being able to extend menus when embedding site editor
This PR adds a simple menu for being able to extend the site editor when
it is used as an embedded application. In particular it adds support for
adding menu items which do not exist in the core site editor framework.
This is useful for instance when say we build a custom version of site
editor which exports lines in a bespoke format. Additionally, while this
approach increases boiler plate, it also allows plugins to be self
contained. This will be useful when scaling site editor up and elad to
more readable self contained code when adding new features.

The approach used is inspired by VSCode's extension API. At startup a
plugin can declare its intent to register a new menu action. When it
does so it will recieve an event handle the plugin should listen for.
The plugin can then listen for the event handle separately thus enabling
a complete integration with the UI of the site editor without much
effort.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>