Early support for undo in `ChangePlugin`
This provides a POC version of undo for the `ChangePlugin`. We still
have to implement undo for movement, adding and deleteing items. One
thing to note is that the current architecture assumes a change at the
individual component level. Also currently, we need `ResMut` access tot
he undo buffer from all change plugins. There may come a time where we may
want to change multiple components in a single operation.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>